backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
httpx
google-generativeai
google-genai